azure-microservice/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
azure-functions==1.17.0
azure-cosmos==4.6.0
